исправил делегат
